% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_functions.R
\name{collapse_sum}
\alias{collapse_sum}
\title{collapse_sum}
\usage{
collapse_sum(x)
}
\arguments{
\item{x}{a numeric or a character vector}
}
\value{
a double equals to the sum of x or a character made by collapsing  unique characters of x
}
\description{
collapse_sum
}
\examples{
collapse_sum(c(3,2,2))
collapse_sum(c("a", "a", "b"))

}
