% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_error.R
\name{evaluate_error}
\alias{evaluate_error}
\title{Evaluate error metrics
evaluate_error}
\usage{
evaluate_error(data, observed = "observed", simulated = "simulated")
}
\arguments{
\item{data}{a dataset with observed and simulated/fitted values}

\item{observed}{name of the column with observed values}

\item{simulated}{name of the column with simulated/fitted values}
}
\value{
a dataset with error metrics
}
\description{
Evaluate error metrics
evaluate_error
}
\examples{
 x <- rnorm(100)
d <- data.frame(observed = x, simulated = x + rnorm(100)  )

evaluate_error(d)

}
