% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n_of.R
\name{sample_n_of}
\alias{sample_n_of}
\title{sample elements from subgroup}
\usage{
sample_n_of(data, size, ...)
}
\arguments{
\item{data}{tibble}

\item{size}{the number  subgroups you want  sample}

\item{...}{the column names  the subgroup you want  sample}
}
\value{
tibble with elements belonging  the sampled subgroups
}
\description{
sample elements from subgroup
}
\examples{
library(dplyr)
 mtcars \%>\% sample_n_of(2,cyl) \%>\% distinct(cyl)

 mtcars \%>\%
 mutate(rowid = 1:n()) \%>\%
 group_by(cyl) \%>\%
 do(sample_n_of(., 2,rowid))
}
