% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_customized.R
\name{which.customized}
\alias{which.customized}
\title{Which customized}
\usage{
which.customized(x, arr.ind = FALSE, useNames = TRUE)
}
\arguments{
\item{x}{a logical vector or array. NAs are allowed and omitted (treated as if FALSE).}

\item{arr.ind}{logical; should array indices be returned when x is an array?}

\item{useNames}{logical indicating if the value of arrayInd() should have (non-null) dimnames at all.}
}
\value{
If TRUE present in x: same thing as base::which, else returns NA
}
\description{
Same function as base::which, except that it returns NA instead of integer(0) when there is not TRUE in x
}
\examples{
which(letters == "a character that doesn't exist")
which.customized(letters == "a character that doesn't exist")
}
