% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_functions.R
\name{collapse_sum}
\alias{collapse_sum}
\title{collapse a vector with a sum or a string depending on vector type
collapse_sum}
\usage{
collapse_sum(x, collapse_character = "_")
}
\arguments{
\item{x}{a numeric or a character vector}

\item{collapse_character}{character to put in the collapse argument of paste function}
}
\value{
a double equals to the sum of x or a character made by collapsing  unique characters of x
}
\description{
collapse a vector with a sum or a string depending on vector type
collapse_sum
}
\examples{
collapse_sum(c(3,2,2))
collapse_sum(c("a", "a", "b"))

}
