% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_functions.R
\name{collapse_unique}
\alias{collapse_unique}
\title{collapse_unique}
\usage{
collapse_unique(x)
}
\arguments{
\item{x}{a numeric or a character vector}
}
\value{
a double equals to the sum of unique values of x or a character made by collapsing unique characters of x
}
\description{
collapse_unique
}
\examples{
collapse_unique(c(3,2,2))
collapse_unique(c("a", "a", "b"))

}
